+++ src/sample.rs
@@ -11,21 +11,27 @@ use std::{
 };
 use tokio::process::Command;
 
-/// Create a 20s sample from `sample_start`.
+/// Create a sample from `sample_start` + `frames`.
 ///
 /// Fast as this uses `-c:v copy`.
 pub async fn copy(
     input: &Path,
     sample_start: Duration,
+    floor_to_sec: bool,
     frames: u32,
     temp_dir: Option<PathBuf>,
 ) -> anyhow::Result<PathBuf> {
+    let mut sample_start_s = sample_start.as_secs_f32();
+    if floor_to_sec {
+        sample_start_s = sample_start_s.floor();
+    }
+
     let mut dest = temporary::process_dir(temp_dir);
     // Always using mkv for the samples works better than, e.g. using mp4 for mp4s
     // see https://github.com/alexheretic/ab-av1/issues/82#issuecomment-1337306325
     dest.push(
         input
-            .with_extension(format!("sample{}+{frames}f.mkv", sample_start.as_secs()))
+            .with_extension(format!("sample{sample_start_s}+{frames}f.mkv"))
             .file_name()
             .unwrap(),
     );
@@ -38,7 +44,7 @@ pub async fn copy(
     // See https://github.com/alexheretic/ab-av1/issues/36#issuecomment-1146634936
     let mut out = Command::new("ffmpeg")
         .arg("-y")
-        .arg2("-ss", sample_start.as_secs().to_string())
+        .arg2("-ss", sample_start_s)
         .arg2("-i", input)
         .arg2("-frames:v", frames)
         .arg2("-c:v", "copy")
@@ -58,7 +64,7 @@ pub async fn copy(
             .arg("-y")
             // try +genpts workaround
             .arg2("-fflags", "+genpts")
-            .arg2("-ss", sample_start.as_secs().to_string())
+            .arg2("-ss", sample_start_s)
             .arg2("-i", input)
             .arg2("-frames:v", frames)
             .arg2("-c:v", "copy")